meeting-tui/pyproject.toml
[project]
name = "meeting-tui"
version = "0.1.0"
description = "A terminal-based meeting companion with live transcription and AI chat"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "textual>=3.0",
    "sounddevice>=0.5",
    "numpy>=1.26",
    "faster-whisper>=1.1",
    "onnxruntime>=1.17",
    "silero-vad>=5.1",
    "httpx>=0.27",
    "openai>=1.50",
    "google-genai>=1.0",
    "rich>=13.0",
    "click>=8.1",
    "python-dotenv>=1.0",
]

[project.scripts]
meeting-tui = "meeting_tui.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/meeting_tui"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
